requirements.txt
# 核心框架
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
python-multipart>=0.0.9
pydantic>=2.9.0
pydantic-settings>=2.5.0

# LangChain 相关
langchain>=0.3.0
langchain-community>=0.3.0
langchain-ollama>=0.2.0

# 向量数据库
chromadb>=0.5.0

# Git 操作
gitpython>=3.1.0

# 基础工具
requests>=2.31.0
python-dotenv>=1.0.0
jinja2>=3.1.0